improvements in reactivity and rendering
`lang.mjs`:

Added the function `reset` which is a counterpart of `deref`.
Intended mainly for observable references from the `obs` module.

`isRec` and associated assertion functions now consider instances of
`Promise` to be non-records. This may help detecting some cases of
forgetting to use `await`.

`isCls` now returns `false` for all functions defined with the keyword
`function`, and `true` for anything defined with the keyword `class`,
as well as built-ins. Not all built-ins have been tested. It produces
false positives for some built-ins, such as `Symbol` and `BigInt`.

`show` now shows own enumerable properties, if any, when describing
functions and classes.

`obs.mjs`:

Dereferencing any observable via `l.deref` or `l.derefAll` is now a
reactive operation.

Added the function `getTar` for non-reactively dereferencing proxies.
This is supported by all types of observables provided by this
package. For inputs which do not specially implement this feature,
this simply returns the input, similarly to `l.deref`.

The new function `l.reset` works on all observables. The support
for `reset` in proxy-based observables is tentative: it requires
the input to be either nil or a record, and patches the target by
copying the own properties of the input, similar to `Object.assign`,
without deleting any pre-existing properties. This may be changed
to a full reset in the future. This is virtually identical to using
`Object.assign`, with the difference that the proxy trap `set` is
invoked only once, and in case of changes, the observable queue is
flushed only once. This may minimize the need for pausing the
synchronous scheduler when reassigning multiple fields.

Added a precaution against some cases of accidental infinite
recurrence. When a callback which runs in an implicitly reactive
context, such in a `recur`, modifies one of the observables which
it also observes, it will not recur from this modification.

`prax.mjs`:

Breaking changes across the board. A redesign of the system.

The renderer no longer has any methods with variadic parameters, and
no longer provides any methods with separate parameters for props and
children. Children have been rolled into props, under the keys `.chi`
or `.children`, which are treated the same. The chi-only methods such
as `.replaceChi` take children as a non-variadic parameter.

Props reactivity can be more fine-grained than before. Individual
props can be functions or observable references. When reactivity is
enabled, which is the default, each such prop will be automatically
updated on changes to any observables it uses, separately from other
props. Replacing a prop deinits the previous recurrent for that prop,
if any.

As a special case, props whose values are functions and keys begin
with `on` are treated non-reactively, assigning event callbacks as
usual.

Like before, the renderer still supports providing a function or an
observable reference as the entire props, or as an individual child
node.

Reverted a recent change where calls to `Ren..mutProps`, or calls to
`Ren..mut` with non-nil props, would entirely replace the previous
props. With children being part of the props, and individual props
having the option to be reactive, the replacement behavior is no
longer ergonomic, and can be a hurdle.

The renderer methods `E` and `S` now allow functions and classes as
the "target", the first argument. They immediately invoke the given
function or class with the props argument in an empty reactive
context, isolated from the reactive context of the caller. Code which
heavily uses observables and reactivity is encouraged to prefer the
calling convention `E(View, props)` over `View(props)` because this
automatically isolates reactive contexts from each other, preventing
accidental over-subscription, which in degenerate cases can lead to
infinite rendering loops without this precaution.

Small breaking change in `PropBui`: removed `.with` and renamed
`.mut`. The previous behavior of `.with` was to adopt the given
object, if any, for subsequent mutations. But remembering where
it was safe to adopt objects, vs where it wasn't and a copy was
needed, was too error prone, and the micro-optimization enabled
by this wasn't measurable anyway. Now it never mutates inputs.

Added `PropBui..chi` which stores `.children`, for compatibility
with the changes above.

`PropBui` no longer converts boolean props to booleans, for
compatibility with passing functions and observable references
in their place.

`dom_reg.mjs`:

Fixed how `.localName` set on superclasses is treated for subclasses.
An inherited "custom" `.localName` is ignored, while an inherited
non-"custom" `.localName` is respected.

`coll.mjs`:

Small breaking change: removed `CompatMap`.

`test.mjs`:

Small breaking change: renamed `optInst` to `instOpt`.

`http_deno.mjs`:

Added the `.txt` content type.

When guessing content types from file extensions, the latter are
treated case-insensitively.